simple disappearing login form with greeting
This commit demonstrates a few important elements.

Foremost, it sets up Rails to render a particular layout for all HTML
requests.  This can be changed so it only applies to certain controller
actions as necessary.  The important part here is that it will allow all
JSON requests to pass through.

The secondary portion of this commit is the SessionCtrl located in
app/assets/javascripts/session.js.  This file defines some simple
actions for loggin a user in (i.e. persisting their nickname) and
determining if the user has logged in.

The layout being rendered (app/views/layouts/angular.html.erb) exhibits
a number of angular directives (e.g ng-show and ng-controller).  This
hooks the markup into the various functions defined by the controller,
showing and hiding elements as necessary.