refactor application to use module
This change highlights some important aspects of the AngularJS module
system.

Firstly, the simple creation thereof, which requires both a name and an
array of its dependencies.

Secondly, a $routeProvider is configured. This service, in AngularJS
terminology, provides the ability to define how client-side routes
interact with controllers and templates.

Thirdly, the view layer of the application has now been split out into
separate templates.  These are scoped within their controller as defined
in the $routeProvider service configuration.

Tangentially, adding an additional path to the asset pipeline is
necessary in order to have the templates available to the client.