add csrf token header to all ajax requests
This step is an essential prerequisite to being able to communicate with
Rails in any meaningful (i.e. not idempotent) way. In doing so, two
important items are exposed.  The former is the #csrf_meta_tag from
Rails, which creates an authenticity token meta tag.  The latter is the
ability to configure the $httpProvider service from the primary
application module.