wire up session negotiation with Rails
With this commit, AngularJS can now fully communicate with the Rails
application.

Using the $resource service from the ngResource module, in combination
with the js-routes gem, it is possible to define a simple RESTful
interface that can send requests and receive responses.

The Session module, dependent upon ngResource, defines this behavior.
As such, it must be injected into the SessionCtrl and the main chatter
application is now dependent upon it.