list of nicknames
With this change, the application is finally beginning to take shape.
Not only can a user set their nickname, they can now see a list of all
the other members of the chat.

In terms of AngularJS features present, this commit exhibits the
behavior of the builtin $timeout service for repeatedly calling a
function for syncing with the server.  Because this causes flicker, the
user list will only update when the length of the array has changed.

The other major feature is the ng-repeat directive which iterates over
the elements in an array and modifies the DOM accordingly.  Underlying
this is the complexity of directives and the two stage compile and
linking process, which need only be glossed over for the present.